Fixes to Validate King
Castling if statement rejects king moves that don't move to file 2 or file 6.  This prevents the king from making a normal move on his first move.  Therefore added 'absolute_file_delta > 1' as if condition so that only castling moves are caught by the castling if statement and normal moves are handled below it.

Removed hypothetical_check of final position as this is done in validate_move.  Added hypothetical check of current position and position where the rook ends (and the king passes through).

Added test that king does not move into his own piece.